Home View changed with live sensor value. O.P.
