+++ app/src/main/java/com/example/safehouse/ui/home/HomeViewModel.java
@@ -1,8 +1,20 @@
 package com.example.safehouse.ui.home;
 
+import android.widget.Toast;
+
+import androidx.annotation.NonNull;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
+
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class HomeViewModel extends ViewModel {
 
@@ -10,7 +22,29 @@ public class HomeViewModel extends ViewModel {
 
     public HomeViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment. The default house of the user will be displayed here.");
+        DatabaseReference mDatabase;
+        mDatabase = FirebaseDatabase.getInstance().getReference();
+
+        mDatabase.child("sensor").addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                List notes = new ArrayList<>();
+                Boolean water_sensor = dataSnapshot.getValue(Boolean.class);
+                if(water_sensor==true){
+                    mText.setValue("Water Sensor is Wet ");
+                }
+                if(water_sensor==false){
+                    mText.setValue("Water Sensor is Dry");
+                }
+
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError databaseError) {
+
+            }
+
+        });
     }
 
     public LiveData<String> getText() {
